feat: create component for showing elasticsearch cluster info
 - fix dashboardPaperCard.tsx component, which now, can have all props from PaperProps
 - default selected source is the first element in the sources array
