+++ src/components/ui/dashboardPaperCard.tsx
@@ -1,11 +1,12 @@
 import Paper, {PaperProps} from "@mui/material/Paper";
 import {useTheme} from "@mui/material/styles";
 
-export default function DashboardPaperCard({sx, children, square = false}: PaperProps) {
+export default function DashboardPaperCard(props: PaperProps) {
     const {palette} = useTheme()
+    const {sx, children, square = false} = props
 
     return (
-        <Paper sx={{bgcolor: palette.background.default, ...sx}} square={square}>
+        <Paper {...props} sx={{bgcolor: palette.background.default, ...sx}} square={square}>
             {children}
         </Paper>
     )
